+++ packages/db/src/queries/audit-logs.ts
@@ -1,5 +1,5 @@
 import { eq, desc, and, gte, sql } from "drizzle-orm";
-import type { Database } from "../client";
+import type { AdminDatabase as Database } from "../d1-client";
 import { auditLogs } from "../schema";
 
 export function auditLogWriteQueries(db: Database) {
@@ -39,7 +39,7 @@ export function auditLogWriteQueries(db: Database) {
 
     async countByActionSince(action: string, since: Date) {
       const rows = await db
-        .select({ count: sql<number>`count(*)::int` })
+        .select({ count: sql<number>`count(*)` })
         .from(auditLogs)
         .where(
           and(eq(auditLogs.action, action), gte(auditLogs.createdAt, since)),
